added authentication to all backend routes, middleware, expiry times, etc

+++ src/routes/module/lessons/block.py
@@ -1,7 +1,10 @@
 import ast
 from flask import request
 from flask_restful import Resource
+from backend.auth import get_jwt_sub
 from backend.database.blocks import BlocksTable
+from backend.database.lessons import LessonsTable
+from backend.database.subscriptions import SubscriptionsTable
 from lib.dataswap.database import SwapDB
 from lib.instilled.instiled import Instil
 from lib.swagdoc.swagdoc import SwagDoc, SwagMethod, SwagParam, SwagResp
@@ -9,7 +12,6 @@ from lib.swagdoc.swagmanager import SwagGen
 
 
 class Block(Resource):
-
     @SwagGen(
         SwagDoc(
             SwagMethod.POST,
@@ -66,6 +68,7 @@ class Block(Resource):
                 ),
             ],
             [SwagResp(200, "Block created"), SwagResp(400, "Block not created")],
+            protected=True
         )
     )
     @Instil("db")
@@ -75,10 +78,21 @@ class Block(Resource):
         block_type = int(request.form.get("block_type", 0))
         order = int(request.form.get("order", 0))
         name = str(request.form.get("name", 0))
-        data = ast.literal_eval(request.form.get("data"))
+
+        if (str_data := request.form.get("data")) is None:
+            return {"message": "Block data not found"}, 400
+        data = ast.literal_eval(str_data)
+
+        if (sub := get_jwt_sub()) is None:
+            return {"message": "Unauthorized"}, 401
+
+        if not LessonsTable.user_can_delete(service, lesson_id, sub):
+            return {"message": "Unauthorized"}, 401
 
         # Try to create a block, if successful return a 200 response
-        if BlocksTable.write_block(service, block_id, lesson_id, block_type, order, name, data):
+        if BlocksTable.write_block(
+            service, block_id, lesson_id, block_type, order, name, data
+        ):
             return {"message": "Block created"}, 200
         # If fails, return a 400 response
         else:
@@ -105,15 +119,22 @@ class Block(Resource):
                     True,
                     "The block id to delete",
                     "1",
-                )
+                ),
             ],
             [SwagResp(200, "Block deleted"), SwagResp(404, "Block not found")],
+            protected=True
         )
     )
     @Instil("db")
     def delete(self, lesson_id: int, service: SwapDB):
         # Get block data from request and convert to integers
         block_id = int(request.form.get("block_id", 0))
+
+        if (sub := get_jwt_sub()) is None:
+            return {"message": "Unauthorized"}, 401
+
+        if not LessonsTable.user_can_delete(service, lesson_id, sub):
+            return {"message": "Unauthorized"}, 401
 
         # Try to delete a block, if successful return a 200 response
         if BlocksTable.delete_block(service, lesson_id, block_id):
@@ -138,16 +159,35 @@ class Block(Resource):
                 )
             ],
             [SwagResp(200, "Returns the lesson sidebar")],
+            protected=True
         )
     )
     @Instil("db")
     def get(self, lesson_id: int, service: SwapDB):
         # Get lesson sidebar with basic blocks using lesson_id
         blocks: list[dict[str, int | str | dict]] = []
-        for block_type, block_id, block_order, block_name, data in BlocksTable.get_blocks(service, lesson_id):
+
+        if (sub := get_jwt_sub()) is None:
+            return {"message": "Unauthorized"}, 401
+
+        if not SubscriptionsTable.can_read_lesson(service, sub, lesson_id):
+            return {"message": "Unauthorized"}, 401
+
+        for (
+            block_type,
+            block_id,
+            block_order,
+            block_name,
+            data,
+        ) in BlocksTable.get_blocks(service, lesson_id):
             blocks.append(
-                {"block_type": block_type, "block_id": block_id, "order": block_order,
-                 "name": block_name, "data": data}
+                {
+                    "block_type": block_type,
+                    "block_id": block_id,
+                    "order": block_order,
+                    "name": block_name,
+                    "data": data,
+                }
             )
             # TODO: implement lesson blocking if not submitted
         return {"blocks": blocks}, 200
